+++ src/openrct2-ui/interface/ViewportInteraction.cpp
@@ -21,7 +21,7 @@
 #include <openrct2/actions/BalloonPressAction.h>
 #include <openrct2/actions/FootpathAdditionRemoveAction.h>
 #include <openrct2/actions/LargeSceneryRemoveAction.h>
-#include <openrct2/actions/ParkEntranceRemoveAction.hpp>
+#include <openrct2/actions/ParkEntranceRemoveAction.h>
 #include <openrct2/actions/SmallSceneryRemoveAction.h>
 #include <openrct2/actions/WallRemoveAction.hpp>
 #include <openrct2/localisation/Localisation.h>
