+++ src/openrct2-ui/windows/Cheats.cpp
@@ -15,7 +15,7 @@
 #include <openrct2/Context.h>
 #include <openrct2/Game.h>
 #include <openrct2/OpenRCT2.h>
-#include <openrct2/actions/ParkSetDateAction.hpp>
+#include <openrct2/actions/ParkSetDateAction.h>
 #include <openrct2/actions/SetCheatAction.hpp>
 #include <openrct2/config/Config.h>
 #include <openrct2/localisation/Date.h>
