+++ src/openrct2-ui/windows/NewCampaign.cpp
@@ -12,7 +12,7 @@
 #include <openrct2-ui/interface/Widget.h>
 #include <openrct2-ui/windows/Window.h>
 #include <openrct2/Game.h>
-#include <openrct2/actions/ParkMarketingAction.hpp>
+#include <openrct2/actions/ParkMarketingAction.h>
 #include <openrct2/config/Config.h>
 #include <openrct2/drawing/Drawing.h>
 #include <openrct2/localisation/Localisation.h>
