+++ src/openrct2-ui/windows/Research.cpp
@@ -11,7 +11,7 @@
 #include <openrct2-ui/interface/Widget.h>
 #include <openrct2-ui/windows/Window.h>
 #include <openrct2/Game.h>
-#include <openrct2/actions/ParkSetResearchFundingAction.hpp>
+#include <openrct2/actions/ParkSetResearchFundingAction.h>
 #include <openrct2/localisation/Localisation.h>
 #include <openrct2/management/Finance.h>
 #include <openrct2/management/NewsItem.h>
