+++ src/openrct2/actions/GameActionRegistration.cpp
@@ -37,13 +37,13 @@
 #include "MazePlaceTrackAction.h"
 #include "MazeSetTrackAction.h"
 #include "NetworkModifyGroupAction.h"
-#include "ParkEntranceRemoveAction.hpp"
-#include "ParkMarketingAction.hpp"
-#include "ParkSetDateAction.hpp"
-#include "ParkSetLoanAction.hpp"
-#include "ParkSetNameAction.hpp"
-#include "ParkSetParameterAction.hpp"
-#include "ParkSetResearchFundingAction.hpp"
+#include "ParkEntranceRemoveAction.h"
+#include "ParkMarketingAction.h"
+#include "ParkSetDateAction.h"
+#include "ParkSetLoanAction.h"
+#include "ParkSetNameAction.h"
+#include "ParkSetParameterAction.h"
+#include "ParkSetResearchFundingAction.h"
 #include "PauseToggleAction.hpp"
 #include "PeepPickupAction.hpp"
 #include "PlaceParkEntranceAction.hpp"
