+++ src/openrct2/actions/ParkSetDateAction.cpp
@@ -7,7 +7,7 @@
  * OpenRCT2 is licensed under the GNU General Public License version 3.
  *****************************************************************************/
 
-#pragma once
+#include "ParkSetDateAction.h"
 
 #include "../Context.h"
 #include "../core/MemoryStream.h"
@@ -16,48 +16,25 @@
 #include "../ui/UiContext.h"
 #include "../ui/WindowManager.h"
 #include "../windows/Intent.h"
-#include "GameAction.h"
 
-DEFINE_GAME_ACTION(ParkSetDateAction, GAME_COMMAND_SET_DATE, GameActions::Result)
+void ParkSetDateAction::Serialise(DataSerialiser& stream)
 {
-private:
-    int32_t _year{};
-    int32_t _month{};
-    int32_t _day{};
+    GameAction::Serialise(stream);
+    stream << DS_TAG(_year) << DS_TAG(_month) << DS_TAG(_day);
+}
 
-public:
-    ParkSetDateAction() = default;
-    ParkSetDateAction(int32_t year, int32_t month, int32_t day)
-        : _year(year)
-        , _month(month)
-        , _day(day)
+GameActions::Result::Ptr ParkSetDateAction::Query() const
+{
+    if (_year <= 0 || _year > MAX_YEAR || _month <= 0 || _month > MONTH_COUNT || _day <= 0 || _day > 31)
     {
+        return MakeResult(GameActions::Status::InvalidParameters, STR_NONE);
     }
 
-    uint16_t GetActionFlags() const override
-    {
-        return GameAction::GetActionFlags() | GameActions::Flags::AllowWhilePaused;
-    }
+    return MakeResult();
+}
 
-    void Serialise(DataSerialiser & stream) override
-    {
-        GameAction::Serialise(stream);
-        stream << DS_TAG(_year) << DS_TAG(_month) << DS_TAG(_day);
-    }
-
-    GameActions::Result::Ptr Query() const override
-    {
-        if (_year <= 0 || _year > MAX_YEAR || _month <= 0 || _month > MONTH_COUNT || _day <= 0 || _day > 31)
-        {
-            return MakeResult(GameActions::Status::InvalidParameters, STR_NONE);
-        }
-
-        return MakeResult();
-    }
-
-    GameActions::Result::Ptr Execute() const override
-    {
-        date_set(_year, _month, _day);
-        return MakeResult();
-    }
-};
+GameActions::Result::Ptr ParkSetDateAction::Execute() const
+{
+    date_set(_year, _month, _day);
+    return MakeResult();
+}
