+++ src/openrct2/libopenrct2.vcxproj
@@ -82,13 +82,13 @@
     <ClInclude Include="actions\MazePlaceTrackAction.h" />
     <ClInclude Include="actions\MazeSetTrackAction.h" />
     <ClInclude Include="actions\NetworkModifyGroupAction.h" />
-    <ClInclude Include="actions\ParkEntranceRemoveAction.hpp" />
-    <ClInclude Include="actions\ParkMarketingAction.hpp" />
-    <ClInclude Include="actions\ParkSetDateAction.hpp" />
-    <ClInclude Include="actions\ParkSetLoanAction.hpp" />
-    <ClInclude Include="actions\ParkSetNameAction.hpp" />
-    <ClInclude Include="actions\ParkSetParameterAction.hpp" />
-    <ClInclude Include="actions\ParkSetResearchFundingAction.hpp" />
+    <ClInclude Include="actions\ParkEntranceRemoveAction.h" />
+    <ClInclude Include="actions\ParkMarketingAction.h" />
+    <ClInclude Include="actions\ParkSetDateAction.h" />
+    <ClInclude Include="actions\ParkSetLoanAction.h" />
+    <ClInclude Include="actions\ParkSetNameAction.h" />
+    <ClInclude Include="actions\ParkSetParameterAction.h" />
+    <ClInclude Include="actions\ParkSetResearchFundingAction.h" />
     <ClInclude Include="actions\PauseToggleAction.hpp" />
     <ClInclude Include="actions\PeepPickupAction.hpp" />
     <ClInclude Include="actions\PlaceParkEntranceAction.hpp" />
